requirements.txt
absl-py>=0.9.0
chex>=0.0.8
dm-tree>=0.1.5
jax>=0.1.55
jaxlib>=0.1.37
numpy>=1.18.0
tf-nightly  # Should be tensorflow>=2.8.0, not yet released as of 2022.01.13.
